requirements.txt
transformers==4.31.0
accelerate==0.21.0
diffusers==0.20.0
huggingface_hub==0.16.4
omegaconf==2.3.0
xformers
IPython==8.14.0
Pillow==9.5.0
matplotlib
ipykernel
einops~=0.4.1
safetensors